Add names to rivers in the Shackles
Rivers in _Isles of the Shackles_ are named in the text rather
than the map. Based on the textual descriptions, add names to:

- Holshod River on Devil's Arches
- Greenscale River and River of Knives on Shark Island
- Bluehook River on Tempest Cay
- Broken Hope River on Raptor Island
- Elsewhere River on Whyrlis Rock

Geometry on some rivers was modified to connect the main river.
Lines were split or joined; no vectors were moved.

Named rivers not added in this commit due to a lack of clarity in
the textual description:

- Libera River on Bag Island
- Ravensfoot River on Motaku Isle